requirements.txt
allure-pytest>=2.9.45
allure-python-commons>=2.9.45
playwright>=1.25.1
pytest>=7.1.2
pytest-forked>=1.4.0
pytest-playwright>=0.3.0
pytest-xdist>=2.5.0